requirements.txt
Flask==3.0.3
Bootstrap-Flask==2.4.0
python-dotenv==1.0.1
requests==2.32.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
Flask-WTF==1.2.1
WTForms==3.1.2
googlemaps==4.10.0